+++ src/LoanActionsHistory.js
@@ -1,4 +1,4 @@
-import _ from 'lodash';
+import { get, upperFirst } from 'lodash';
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
 import Link from 'react-router-dom/Link';
@@ -37,13 +37,26 @@ class LoanActionsHistory extends React.Component {
       resourceShouldRefresh: true,
       path: 'users?query=(%{userIds.query})',
     },
+    requests: {
+      type: 'okapi',
+      path: 'circulation/requests',
+      resourceShouldRefresh: true,
+      records: 'requests',
+      accumulate: 'true',
+      fetch: false,
+    },
     loanActions: {
       type: 'okapi',
       records: 'loans',
       resourceShouldRefresh: true,
       GET: {
-        path: 'loan-storage/loan-history?query=(id==!{loanid})&timestamp=%{timestamp.time}',
+        path: 'loan-storage/loan-history?query=(id==!{loanid})&timestamp=%{timestamp.time}&limit=100',
       },
+    },
+    loanPolicies: {
+      type: 'okapi',
+      records: 'loanPolicies',
+      path: 'loan-policy-storage/loan-policies?query=(id==!{loan.loanPolicyId})',
     },
     loanAccountsActions: {
       type: 'okapi',
@@ -63,6 +76,10 @@ class LoanActionsHistory extends React.Component {
     mutator: PropTypes.shape({
       loanActionsWithUser: PropTypes.shape({
         replace: PropTypes.func,
+      }),
+      requests: PropTypes.shape({
+        GET: PropTypes.func,
+        reset: PropTypes.func,
       }),
       userIds: PropTypes.shape({
         replace: PropTypes.func,
@@ -97,12 +114,14 @@ class LoanActionsHistory extends React.Component {
     this.showTitle = this.showTitle.bind(this);
     this.showChangeDueDateDialog = this.showChangeDueDateDialog.bind(this);
     this.hideChangeDueDateDialog = this.hideChangeDueDateDialog.bind(this);
+    this.getOpenRequestsCount = this.getOpenRequestsCount.bind(this);
 
     this.state = {
       loanActionCount: 0,
       nonRenewedLoanItems: [],
       nonRenewedLoansModalOpen: false,
       changeDueDateDialogOpen: false,
+      requestsCount: {},
     };
   }
 
@@ -135,17 +154,22 @@ class LoanActionsHistory extends React.Component {
     return null;
   }
 
+  componentDidMount() {
+    this.getOpenRequestsCount();
+  }
+
   componentDidUpdate(prevProps, prevState) {
     const { loan, resources: { loanActions, users } } = this.props;
 
     if (this.state.loanActionCount && this.state.loanActionCount !== prevState.loanActionCount) {
       this.joinLoanActionsWithUser(loanActions.records, users.records, loan);
     }
+    if (this.props.loan.itemId !== prevProps.loan.itemId) this.getOpenRequestsCount();
   }
 
   getContributorslist(loan) {
     this.loan = loan;
-    const contributors = _.get(this.loan, ['item', 'contributors']);
+    const contributors = get(this.loan, ['item', 'contributors']);
     const contributorsList = [];
     if (typeof contributors !== 'undefined') {
       Object.keys(contributors).forEach(contributor => contributorsList.push(`${contributors[contributor].name}; `));
@@ -184,16 +208,38 @@ class LoanActionsHistory extends React.Component {
     const { loan, user } = this.props;
     const promise = this.props.renew(loan, user);
     const singleRenewalFailure = [];
+
     promise
-      .then(() => this.showCallout())
-      .catch(() => {
-        singleRenewalFailure.push(loan);
-      });
+      .then(() => this.onRenew())
+      .catch(() => singleRenewalFailure.push(loan));
+
     return promise;
   }
 
+  onRenew() {
+    this.showCallout();
+    this.refreshLoanActions();
+  }
+
+  refreshLoanActions() {
+    this.props.mutator.timestamp.replace({ time: Date.now() });
+  }
+
+  getOpenRequestsCount() {
+    const q = `itemId==${this.props.loan.itemId}`;
+    const query = `(${q}) and status==("Open - Awaiting pickup" or "Open - Not yet filled") sortby requestDate desc`;
+    this.props.mutator.requests.reset();
+    this.props.mutator.requests.GET({ params: { query } }).then((requestRecords) => {
+      const requestCountObject = requestRecords.reduce((map, record) => {
+        map[record.itemId] = map[record.itemId] ? ++map[record.itemId] : 1;
+        return map;
+      }, {});
+      this.setState({ requestsCount: requestCountObject });
+    });
+  }
+
   getFeeFine() {
-    const accounts = _.get(this.props.resources, ['loanAccountsActions', 'records'], []);
+    const accounts = get(this.props.resources, ['loanAccountsActions', 'records'], []);
     let amount = 0;
     accounts.forEach(a => {
       amount += parseFloat(a.amount);
@@ -203,7 +249,7 @@ class LoanActionsHistory extends React.Component {
 
   feefinedetails = (e) => {
     const loan = this.loan || {};
-    const accounts = _.get(this.props.resources, ['loanAccountsActions', 'records'], []);
+    const accounts = get(this.props.resources, ['loanAccountsActions', 'records'], []);
     if (accounts.length === 1) {
       this.props.onClickViewAccountActionsHistory(e, { id: accounts[0].id });
     } else if (accounts.length > 1) {
@@ -259,14 +305,14 @@ class LoanActionsHistory extends React.Component {
 
   showTitle(loan) {
     this.loan = loan;
-    const title = `${_.get(this.loan, ['item', 'title'], '')}`;
+    const title = `${get(this.loan, ['item', 'title'], '')}`;
     if (title) {
       const titleTodisplay = (title.length >= 77) ? `${title.substring(0, 77)}...` : title;
       return <KeyValue
         label={this.props.stripes.intl.formatMessage({ id: 'ui-users.loans.columns.title' })}
         value={(
-          <Link to={`/inventory/view/${_.get(this.loan, ['item', 'instanceId'], '')}`}>
-            {`${titleTodisplay} (${_.get(this.loan, ['item', 'materialType', 'name'])})`}
+          <Link to={`/inventory/view/${get(this.loan, ['item', 'instanceId'], '')}`}>
+            {`${titleTodisplay} (${get(this.loan, ['item', 'materialType', 'name'])})`}
           </Link>
         )}
       />;
@@ -287,20 +333,23 @@ class LoanActionsHistory extends React.Component {
   }
 
   render() {
-    const { onCancel, loan, patronGroup, user, resources: { loanActionsWithUser }, stripes: { intl } } = this.props;
+    const { onCancel, loan, patronGroup, user, resources: { loanActionsWithUser, loanPolicies }, stripes: { intl } } = this.props;
     const { nonRenewedLoanItems } = this.state;
     const loanActionsFormatter = {
       action: la => intl.formatMessage({ id: loanActionMap[la.action] }),
-      actionDate: la => this.formatDateTime(_.get(la, ['metadata', 'updatedDate'], '-')),
+      actionDate: la => this.formatDateTime(get(la, ['metadata', 'updatedDate'], '-')),
       dueDate: la => this.formatDateTime(la.dueDate),
       itemStatus: la => la.itemStatus,
       Source: la => <Link to={`/users/view/${la.user.id}`}>{getFullName(la.user)}</Link>,
     };
 
+    const loanPolicyName = get(loanPolicies, 'records[0].name', '-');
+    const requestCount = this.state.requestsCount[this.props.loan.itemId];
+    const requestQueueValue = requestCount ? <Link to={`/requests?filters=requestStatus.open%20-%20not%20yet%20filled%2CrequestStatus.open%20-%20awaiting%20pickup&query=${this.props.loan.item.barcode}&sort=Request%20Date`}>{requestCount}</Link> : 0;
     const contributorsList = this.getContributorslist(loan);
     const contributorsListString = contributorsList.join(' ');
     const contributorsLength = contributorsListString.length;
-    const loanStatus = _.get(loan, ['status', 'name'], '-');
+    const loanStatus = get(loan, ['status', 'name'], '-');
     const buttonDisabled = (loanStatus && loanStatus === 'Closed');
     // Number of characters to trucate the string = 77
     const listTodisplay = (contributorsList === '-') ? '-' : (contributorsListString.length >= 77) ? `${contributorsListString.substring(0, 77)}...` : `${contributorsListString.substring(0, contributorsListString.length - 2)}`;
@@ -320,7 +369,7 @@ class LoanActionsHistory extends React.Component {
 
     return (
       <Paneset isRoot>
-        <Pane id="pane-loandetails" defaultWidth="100%" dismissible onClose={onCancel} paneTitle={`${intl.formatMessage({ id: 'ui-users.loans.loanDetails' })} - ${getFullName(user)} (${_.upperFirst(patronGroup.group)})`}>
+        <Pane id="pane-loandetails" defaultWidth="100%" dismissible onClose={onCancel} paneTitle={`${intl.formatMessage({ id: 'ui-users.loans.loanDetails' })} - ${getFullName(user)} (${upperFirst(patronGroup.group)})`}>
           <Row>
             <span>
               <Button disabled={buttonDisabled} buttonStyle="primary" onClick={this.renew}>{this.props.stripes.intl.formatMessage({ id: 'ui-users.renew' })}</Button>
@@ -343,18 +392,18 @@ class LoanActionsHistory extends React.Component {
               {this.showContributors(contributorsList, listTodisplay, contributorsLength)}
             </Col>
             <Col xs={2}>
-              <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.columns.barcode' })} value={<Link to={`/inventory/view/${_.get(loan, ['item', 'instanceId'], '')}/${_.get(loan, ['item', 'holdingsRecordId'], '')}/${_.get(loan, ['itemId'], '')}`}>{_.get(loan, ['item', 'barcode'], '')}</Link>} />
+              <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.columns.barcode' })} value={<Link to={`/inventory/view/${get(loan, ['item', 'instanceId'], '')}/${get(loan, ['item', 'holdingsRecordId'], '')}/${get(loan, ['itemId'], '')}`}>{get(loan, ['item', 'barcode'], '')}</Link>} />
             </Col>
             <Col xs={2}>
-              <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.details.callNumber' })} value={_.get(loan, ['item', 'callNumber'], '-')} />
+              <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.details.callNumber' })} value={get(loan, ['item', 'callNumber'], '-')} />
             </Col>
             <Col xs={2}>
-              <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.details.location' })} value={_.get(loan, ['item', 'location', 'name'], '-')} />
+              <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.details.location' })} value={get(loan, ['item', 'location', 'name'], '-')} />
             </Col>
           </Row>
           <Row>
             <Col xs={2}>
-              <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.columns.itemStatus' })} value={_.get(loan, ['item', 'status', 'name'], '-')} />
+              <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.columns.itemStatus' })} value={get(loan, ['item', 'status', 'name'], '-')} />
             </Col>
             <Col xs={2}>
               <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.columns.dueDate' })} value={this.formatDateTime(loan.dueDate) || '-'} />
@@ -363,7 +412,7 @@ class LoanActionsHistory extends React.Component {
               <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.columns.returnDate' })} value={this.formatDateTime(loan.returnDate) || '-'} />
             </Col>
             <Col xs={2}>
-              <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.details.renewalCount' })} value={_.get(loan, ['renewalCount'], '-')} />
+              <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.details.renewalCount' })} value={get(loan, ['renewalCount'], '-')} />
             </Col>
             <Col xs={2}>
               <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.details.claimedReturned' })} value="TODO" />
@@ -371,7 +420,7 @@ class LoanActionsHistory extends React.Component {
           </Row>
           <Row>
             <Col xs={2}>
-              <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.details.loanPolicy' })} value={<Link to={`/settings/circulation/loan-policies/${loan.loanPolicyId}`}>TODO</Link>} />
+              <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.details.loanPolicy' })} value={<Link to={`/settings/circulation/loan-policies/${loan.loanPolicyId}`}>{loanPolicyName}</Link>} />
             </Col>
             <Col xs={2}>
               <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.columns.loanDate' })} value={this.formatDateTime(loan.loanDate) || '-'} />
@@ -385,7 +434,7 @@ class LoanActionsHistory extends React.Component {
               />
             </Col>
             <Col xs={2}>
-              <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.details.requestQueue' })} value="TODO" />
+              <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.details.requestQueue' })} value={requestQueueValue} />
             </Col>
             <Col xs={2}>
               <KeyValue label={intl.formatMessage({ id: 'ui-users.loans.details.lost' })} value="TODO" />