+++ src/components/ProxyGroup/ProxyEditList/ProxyEditList.js
@@ -114,17 +114,32 @@ export default class ProxyEditList extends React.Component {
           <Col xs>
             <h3 className={css.label}>{label}</h3>
           </Col>
+        </Row>
+        {items.length ?
+          items :
+          <p className={css.isEmptyMessage}>
+            <FormattedMessage
+              id="ui-users.noItemFound"
+              values={{ item: name }}
+            />
+          </p>
+        }
+        <Row>
           <Col xs={4}>
-            <Layout className="right">
+            <Layout>
               <Pluggable
                 aria-haspopup="true"
                 type="find-user"
                 {...this.props}
                 dataKey={name}
-                searchLabel={intl.formatMessage({ id: 'stripes-components.addNew' })}
+                searchLabel={
+                  name === 'proxies' ?
+                    <FormattedMessage id="ui-users.sponsor.addSponsor" /> :
+                    <FormattedMessage id="ui-users.proxy.addProxy" />
+                }
                 searchButtonStyle="default"
                 selectUser={user => this.onAdd(user)}
-                visibleColumns={['active', 'name', 'patronGroup', 'username', 'barcode']}
+                visibleColumns={['status', 'name', 'patronGroup', 'username', 'barcode']}
                 columnMapping={columnMapping}
                 disableRecordCreation={disableRecordCreation}
               >
@@ -133,7 +148,6 @@ export default class ProxyEditList extends React.Component {
             </Layout>
           </Col>
         </Row>
-        {items.length ? items : <p className={css.isEmptyMessage}><FormattedMessage id="ui-users.noItemFound" values={{ item: name }} /></p>}
       </div>
     );
   }